requirements.txt
fastapi==0.115.14
python-dotenv==1.1.1
python-multipart==0.0.20
supabase==2.16.0
uvicorn==0.34.3
email-validator==2.2.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
